refactor: extract floatterminal and markdown-preview to lazy modules
- Move floatterminal config from init.lua to lua/floatterminal/init.lua
  - Migrate toggleterm.nvim setup with terminal keymaps
  - Configures ESC to exit terminal mode
  - Maps Ctrl+hjkl for window navigation from terminal

- Move markdown-preview config from init.lua to lua/markdown-preview/init.lua
  - Migrate mdpreview.nvim with nvim-terminal.lua dependency
  - Lazy-loads on markdown filetype

- Update init.lua to use lazy.nvim import pattern
  - Replaces inline plugin specs with module imports
  - Follows established module structure for consistency

- Add comprehensive documentation to README.md
  - Document all plugins with descriptions and links
  - Add library structure with module pattern explanation
  - List supported LSP languages (C/C++, C#, TypeScript, Python, Lua, Bash)
  - Add keybindings reference table
